Move lint machinery into a separate file
